added tests for user service

+++ src/main/java/org/sbezgin/p2016/db/dao/FileDAO.java
@@ -7,25 +7,25 @@ import org.sbezgin.p2016.db.entity.file.Folder;
 import java.util.List;
 
 public interface FileDAO {
-    AbstractFile getFileByID(int userID, long fileID);
+    AbstractFile getFileByID(Long userID, long fileID);
 
-    List<AbstractFile> getFileByIDs(int userID, List<Long> fileIDs);
+    List<AbstractFile> getFileByIDs(Long userID, List<Long> fileIDs);
 
-    List<AbstractFile> getFilesByName(int userID, String folderPath, String fileName);
+    List<AbstractFile> getFilesByName(Long userID, String folderPath, String fileName);
 
-    Folder getFolder(int userID, long folderID);
+    Folder getFolder(Long userID, long folderID);
 
-    void saveOrUpdateFile(int userID, AbstractFile file);
+    void saveOrUpdateFile(Long userID, AbstractFile file);
 
-    void saveOrUpdateFiles(int userID, List<AbstractFile> files);
+    void saveOrUpdateFiles(Long userID, List<AbstractFile> files);
 
-    int deleteFile(int userID, long fileID);
+    int deleteFile(Long userID, long fileID);
 
-    List<AbstractFile> getRootFiles(int ownerID);
+    List<AbstractFile> getRootFiles(Long ownerID);
 
-    List<AbstractFile> getChildren(int userID, long folderID, int start, int end);
+    List<AbstractFile> getChildren(Long userID, long folderID, int start, int end);
 
-    List<AbstractFile> getFilesByIDs(int userID, List<Long> idList);
+    List<AbstractFile> getFilesByIDs(Long userID, List<Long> idList);
 
-    List<AbstractFile> getAllChildren(int currentUserId, long fileID);
+    List<AbstractFile> getAllChildren(Long currentUserId, long fileID);
 }
